Moves udp hole punching logic into hole_punching module
